requirements.txt
zhipuai
chromadb
pydantic>=2.7.0
click
rich
python-dotenv
sniffio
anyio
